+++ test/project_types/extension/commands/push_test.rb
@@ -5,6 +5,7 @@ require "project_types/extension/extension_test_helpers"
 module Extension
   module Commands
     class PushTest < MiniTest::Test
+      include TestHelpers::FakeUI
       include ExtensionTestHelpers::Messages
 
       def setup